sql: remove SimplifyExpr
This is dead code now.

Release note: None
